requirement.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
langchain-text-splitters
sentence_transformers
langchain_huggingface
langchain_chroma
groq
langchain_groq
pymupdf
langsmith
streamlit
fastapi
uvicorn
wikipedia
arxiv
faiss-cpu
pdfplumber
mysql-connector-python
SQLAlchemy
pymysql
validators
youtube_transcript_api
unstructured
pytube
numexpr
langgraph-checkpoint-sqlite
langgraph
langchain
langchain_community
ddgs
sentence-transformers
langchain_chroma
chromadb
langchain_huggingface